:recycle: migrate tests about AlignmentWrigintModifierTests
